Lots of small changes, start of work to be multi-database type compatible

+++ Staat/Models/MonitorData.cs
@@ -24,7 +24,7 @@ namespace Staat.Models
     public class MonitorData : ITimeStampedModel
     {
         [Key] public int Id { get; set; }
-        public int PingTime { get; set; }
+        public long PingTime { get; set; }
         public bool Available { get; set; }
         public bool? SslValid { get; set; }
         public string FailureReason { get; set; }
